+++ src/lib/util/__mocks__/hooks/mockUseMiniCart.ts
@@ -144,7 +144,6 @@ const mock: TUseMiniCart = {
     },
     currencyCode: CurrencyEnum.Eur,
     handleBeginEditItem: () => {},
-    handleDismiss: () => {},
     handleEndEditItem: () => {},
     handleClose: () => {},
     handleToggleCart: () => {},
