--- src/peregrine/lib/talons/adeoweb/Customer/useDeleteCustomerAddress.d.ts
@@ -1,14 +1,0 @@
-import { DocumentNode } from 'graphql';
-
-type TUseDeleteCustomerAddressProps = {
-    deleteCustomerAddressMutation: DocumentNode;
-};
-
-type TUseDeleteCustomerAddress = {
-    isDeletingAddress: boolean;
-    deleteCustomerAddress: (id: number) => void;
-};
-
-export function useDeleteCustomerAddress(
-    props: TUseDeleteCustomerAddressProps
-): TUseDeleteCustomerAddress;
