+++ src/peregrine/lib/talons/adeoweb/Customer/useDeleteCustomerAddress.ts
@@ -1,10 +1,23 @@
+import { DocumentNode } from 'graphql';
+
 import { useMutation } from '@apollo/react-hooks';
 import { useCallback } from 'react';
 
 import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
 import { fetchPolicy } from 'src/peregrine/lib/util/adeoweb/fetchPolicy';
 
-export const useDeleteCustomerAddress = ({ deleteCustomerAddressMutation }) => {
+type TUseDeleteCustomerAddressProps = {
+    deleteCustomerAddressMutation: DocumentNode;
+};
+
+type TUseDeleteCustomerAddress = {
+    isDeletingAddress: boolean;
+    deleteCustomerAddress: (id: number) => void;
+};
+
+export const useDeleteCustomerAddress = ({
+    deleteCustomerAddressMutation
+}: TUseDeleteCustomerAddressProps): TUseDeleteCustomerAddress => {
     const [
         { isDeletingAddress },
         { deleteCustomerAddress: deleteCustomerAddressAction }
