+++ src/peregrine/lib/talons/adeoweb/Customer/useSetDefaultCustomerAddress.ts
@@ -1,3 +1,5 @@
+import { DocumentNode } from 'graphql';
+
 import { useMutation } from '@apollo/react-hooks';
 import { useCallback } from 'react';
 
@@ -6,12 +8,24 @@ import { useAwaitQuery } from '@magento/peregrine/lib/hooks/useAwaitQuery';
 import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
 import { fetchPolicy } from 'src/peregrine/lib/util/adeoweb/fetchPolicy';
 
+type TUseSetDefaultCustomerAddressProps = {
+    updateCustomerAddressMutation: DocumentNode;
+    getCustomerQuery: DocumentNode;
+    shipping?: boolean;
+    billing?: boolean;
+};
+
+type TUseSetDefaultCustomerAddress = {
+    isUpdatingAddress: boolean;
+    setCustomerDefaultAddress: (id: number) => void;
+};
+
 export const useSetDefaultCustomerAddress = ({
     updateCustomerAddressMutation,
     getCustomerQuery,
     shipping,
     billing
-}) => {
+}: TUseSetDefaultCustomerAddressProps): TUseSetDefaultCustomerAddress => {
     const [{ isUpdatingAddress }, { setDefaultAddress }] = useUserContext();
     const fetchUserDetails = useAwaitQuery(getCustomerQuery);
     const [updateCustomerAddressQuery] = useMutation(
