--- src/peregrine/lib/talons/adeoweb/LayoutModes/useLayoutModes.d.ts
@@ -1,7 +1,0 @@
-export type TUseLayoutModes = {
-    layoutMode: string;
-    setGrid: () => void;
-    setList: () => void;
-};
-
-export function useLayoutModes(): TUseLayoutModes;
