+++ src/peregrine/lib/talons/adeoweb/Logo/useLogo.ts
@@ -1,6 +1,18 @@
 import { useConfigContext } from '../../../context/adeoweb/config';
 
-export const useLogo = props => {
+type TUseLogoProps = {
+    logoPathPrefix?: string;
+};
+
+export type TUseLogo = {
+    storeName: string;
+    logoSrc: string;
+    logoAlt: string;
+    logoHeight?: number;
+    logoWidth?: number;
+};
+
+export const useLogo = (props: TUseLogoProps): TUseLogo => {
     const { logoPathPrefix } = props;
     const [
         {
@@ -14,10 +26,10 @@ export const useLogo = props => {
     const logoSrc = [logoPathPrefix, headerLogoSrc].join('');
 
     return {
-        storeName: storeName || '',
+        storeName: storeName ?? '',
         logoSrc,
-        logoAlt: logoAlt || '',
-        logoHeight: logoHeight || undefined,
-        logoWidth: logoWidth || undefined
+        logoAlt: logoAlt ?? '',
+        logoHeight: logoHeight ?? undefined,
+        logoWidth: logoWidth ?? undefined
     };
 };
