+++ src/peregrine/lib/talons/adeoweb/Menu/useMenu.ts
@@ -1,11 +1,22 @@
+import { DocumentNode } from 'graphql';
+
 import { useEffect } from 'react';
 
 import { useAwaitQuery } from '@magento/peregrine/lib/hooks/useAwaitQuery';
-import { getSearchParam } from '@magento/peregrine/lib/hooks/useSearchParam';
 
+import { TCategoryTree } from 'src/lib/types/graphql/Category';
 import { useMenuContext } from 'src/peregrine/lib/context/adeoweb/menu';
 
-export const useMenu = ({ getMenuQuery }) => {
+type TUseMenuProps = {
+    getMenuQuery: DocumentNode;
+};
+
+export type TUseMenu = {
+    error: null | Error;
+    children: TCategoryTree[];
+};
+
+export const useMenu = ({ getMenuQuery }: TUseMenuProps): TUseMenu => {
     const [{ loadingError, children }, { getMenu }] = useMenuContext();
 
     const fetchMenu = useAwaitQuery(getMenuQuery);
