--- src/peregrine/lib/util/adeoweb/remapToCustomerAddressFormValues.d.ts
@@ -1,6 +1,0 @@
-import { TCustomerAddressFormValues } from 'src/lib/components/Customer';
-import { TCustomerAddress } from 'src/lib/types/graphql/Customer';
-
-export default function remapToCustomerAddressFormValues(
-    address: TCustomerAddress
-): TCustomerAddressFormValues;
